Tweak constants to make simulation more accurate

+++ src/main/java/somephysicsthing/solarsystem/propertytraits/HasMass.java
@@ -4,5 +4,5 @@ package somephysicsthing.solarsystem.propertytraits;
  * Something that has mass
  */
 public interface HasMass {
-    float getMass();
+    double getMass();
 }
